anunnakis_roteiros/requirements.txt
pyside6
google-genai
requests
beautifulsoup4
Pillow
sqlite-utils
python-dotenv
icrawler
